+++ app.rb
@@ -1,321 +1,20 @@
 # frozen_string_literal: true
 
-require 'fileutils'
-require 'json'
-require 'netomox'
-require_relative 'lib/rest_api_base'
-require_relative 'model_defs/topology_builder/networks'
-require_relative 'lib/convert_namespace/namespace_converter'
-require_relative 'lib/convert_namespace/layer_filter'
-require_relative 'lib/convert_topology/batfish_converter'
-require_relative 'lib/convert_topology/containerlab_converter'
+$LOAD_PATH.unshift File.dirname(__FILE__)
 
-# Directory to save batfish query answers
-QUERIES_DIR = ENV.fetch('MDDO_QUERIES_DIR', 'queries')
-# Directory to save topology json from batfish query answers
-TOPOLOGIES_DIR = ENV.fetch('MDDO_TOPOLOGIES_DIR', 'topologies')
-# (temporary) layout file directory
-MODEL_DEFS_DIR = './model_defs'
+require_relative 'lib/api/rest_api_base'
+require_relative 'lib/api/topologies'
 
 module NetomoxExp
-  # rubocop:disable Metrics/ClassLength
+  # Directory to save batfish query answers
+  QUERIES_DIR = ENV.fetch('MDDO_QUERIES_DIR', 'queries')
+  # Directory to save topology json from batfish query answers
+  TOPOLOGIES_DIR = ENV.fetch('MDDO_TOPOLOGIES_DIR', 'topologies')
+  # (temporary) layout file directory
+  MODEL_DEFS_DIR = './model_defs'
 
   # Netomox REST API definition
   class NetomoxRestApi < RestApiBase
-    # rubocop:disable Metrics/BlockLength
-    helpers do
-      # @param [String] file_path File path to read
-      # @return [Object]
-      def read_json_file(file_path)
-        error!("Not found: topology file: #{file_path}", 404) unless File.exist?(file_path)
-
-        JSON.parse(File.read(file_path))
-      end
-
-      # @param [String] file_path File path to save
-      # @param [void]
-      def save_json_file(file_path, data)
-        logger.warn "[save_json_file] path=#{file_path}"
-        FileUtils.mkdir_p(File.dirname(file_path))
-        File.open(file_path, 'w') { |file| JSON.dump(data, file) }
-      end
-
-      # @param [String] network Network name
-      # @param [String] snapshot Snapshot name
-      # @return [Object] Networks data
-      def read_topology_file(network, snapshot)
-        topology_file = File.join(TOPOLOGIES_DIR, network, snapshot, 'topology.json')
-        read_json_file(topology_file)
-      end
-
-      # @param [String] network Network name
-      # @param [String] snapshot Snapshot name
-      # @return [Netomox::Topology::Networks] topology instance
-      def read_topology_instance(network, snapshot)
-        topology_data = read_topology_file(network, snapshot)
-        Netomox::Topology::Networks.new(topology_data)
-      end
-
-      # @param [String] network Network name
-      # @return [String] file path
-      def ns_convert_table_file(network)
-        File.join(TOPOLOGIES_DIR, network, 'ns_convert_table.json')
-      end
-
-      # @param [String] network Network name
-      # @return [void]
-      def save_ns_convert_table(network, data)
-        save_json_file(ns_convert_table_file(network), data)
-      end
-
-      # @param [String] network Network name
-      # @return [Hash] convert_table
-      def read_ns_convert_table(network)
-        read_json_file(ns_convert_table_file(network))
-      end
-
-      # @param [String] network Network name
-      # @return [NamespaceConverter] Namespace converter without topology data
-      def ns_converter_wo_topology(network)
-        ns_converter = NamespaceConverter.new
-        ns_converter.reload_convert_table(read_ns_convert_table(network))
-        ns_converter
-      end
-    end
-    # rubocop:enable Metrics/BlockLength
-
-    # rubocop:disable Metrics/BlockLength
-    namespace 'topologies' do
-      desc 'Post (register) netoviz index'
-      params do
-        requires :index_data, type: Array, desc: 'List of topology'
-      end
-      post 'index' do
-        index_file = File.join(TOPOLOGIES_DIR, '_index.json')
-        save_json_file(index_file, params[:index_data])
-        # response
-        {}
-      end
-
-      params do
-        requires :network, type: String, desc: 'Network name'
-      end
-
-      resource ':network' do
-        desc 'Delete topologies data'
-        delete do
-          network_dir = File.join(TOPOLOGIES_DIR, params[:network])
-          FileUtils.rm_rf(network_dir)
-          # response
-          ''
-        end
-
-        resource 'ns_convert_table' do
-          desc 'Post convert_table'
-          params do
-            optional :origin_snapshot, type: String, desc: 'Origin snapshot name'
-            optional :convert_table, type: Hash, desc: 'Convert table'
-            mutually_exclusive :origin_snapshot, :convert_table, message: 'are exclusive cannot pass both params'
-          end
-          post do
-            network = params[:network]
-            ns_converter = NamespaceConverter.new
-            if params.key?(:origin_snapshot)
-              snapshot = params[:origin_snapshot]
-              logger.info "Initialize namespace convert table with snapshot: #{network}/#{snapshot}"
-              ns_converter.make_convert_table(read_topology_file(network, snapshot))
-            else
-              logger.info "Update namespace convert table of network: #{network}"
-              ns_converter.reload_convert_table(params[:convert_table])
-            end
-            save_ns_convert_table(network, ns_converter.convert_table)
-            # response
-            {}
-          end
-
-          desc 'Get convert_table'
-          get do
-            # response
-            read_ns_convert_table(params[:network])
-          end
-
-          desc 'Delete convert_table'
-          delete do
-            FileUtils.rm_f(ns_convert_table_file(params[:network]))
-            # response
-            ''
-          end
-
-          desc 'Convert hostname'
-          params do
-            requires :host_name, type: String, desc: 'Host name to convert'
-            optional :if_name, type: String, desc: 'Interface name to convert'
-          end
-          post 'query' do
-            network, host_name = %i[network host_name].map { |key| params[key] }
-            ns_converter = ns_converter_wo_topology(network)
-            begin
-              resp = { origin_host: host_name, target_host: ns_converter.node_name_table.convert(host_name) }
-              if params.key?(:if_name)
-                if_name = params[:if_name]
-                resp[:origin_if] = if_name
-                resp[:target_if] = ns_converter.tp_name_table.convert(host_name, if_name)
-              end
-              # response
-              resp
-            rescue StandardError
-              error!("#{params} not found in convert table", 404)
-            end
-          end
-        end
-
-        params do
-          requires :snapshot, type: String, desc: 'Snapshot name'
-        end
-
-        resource ':snapshot' do
-          desc 'Post (register) topology data'
-          resource 'topology' do
-            params do
-              optional :topology_data, type: Hash, desc: 'RFC8345 topology data'
-            end
-            post do
-              network, snapshot = %i[network snapshot].map { |key| params[key] }
-              api_path = "/topologies/#{network}/#{snapshot}/topology"
-
-              topology_data =
-                if params.key?(:topology_data)
-                  logger.debug("[post #{api_path}] posted topology data")
-                  params[:topology_data]
-                else
-                  query_snapshot_dir = File.join(QUERIES_DIR, network, snapshot)
-                  logger.debug("[post #{api_path}] query_snapshot_dir: #{query_snapshot_dir}")
-                  TopologyBuilder.generate_data(query_snapshot_dir)
-                end
-
-              # generate(overwrite) topology data
-              topology_dir = File.join(TOPOLOGIES_DIR, network, snapshot)
-              topology_file = File.join(topology_dir, 'topology.json')
-              save_json_file(topology_file, topology_data)
-
-              # copy layout file if found
-              layout_file = File.join(MODEL_DEFS_DIR, network, snapshot, 'layout.json')
-              FileUtils.cp(layout_file, File.join(topology_dir, 'layout.json')) if File.exist?(layout_file)
-            end
-
-            desc 'Get topology data'
-            get do
-              network, snapshot = %i[network snapshot].map { |key| params[key] }
-              # response
-              read_topology_file(network, snapshot)
-            end
-
-            desc 'Get topology data (L3+ layers)'
-            get 'upper_layer3' do
-              network, snapshot = %i[network snapshot].map { |key| params[key] }
-              layer_filter = LayerFilter.new(read_topology_file(network, snapshot))
-              # response
-              layer_filter.filter
-            end
-
-            params do
-              requires :layer, type: String, desc: 'Network layer'
-            end
-            resource ':layer' do
-              desc 'convert layer data to batfish layer1_topology.json'
-              get 'batfish_layer1_topology' do
-                network, snapshot, layer = %i[network snapshot layer].map { |key| params[key] }
-                topology_data = read_topology_file(network, snapshot)
-                ns_converter = ns_converter_wo_topology(network)
-                bf_converter = BatfishConverter.new(topology_data, layer, ns_converter)
-                # response
-                bf_converter.convert
-              end
-
-              desc 'convert layer data to container-lab topology json'
-              params do
-                optional :env_name, type: String, desc: 'Environment name (for container-lab)', default: 'emulated'
-              end
-              get 'containerlab_topology' do
-                network, snapshot, layer, env_name = %i[network snapshot layer env_name].map { |key| params[key] }
-                topology_data = read_topology_file(network, snapshot)
-                ns_converter = ns_converter_wo_topology(network)
-                clab_converter = ContainerLabConverter.new(topology_data, layer, ns_converter, { env_name: })
-                # response
-                clab_converter.convert
-              end
-
-              desc 'Get all nodes in the layer'
-              params do
-                optional :node_type, type: String, desc: 'Node type (segment/node/endpoint)'
-              end
-              get 'nodes' do
-                network, snapshot, layer = %i[network snapshot layer].map { |key| params[key] }
-                nws = read_topology_instance(network, snapshot)
-                nw = nws.find_network(layer)
-                error!("#{network}/#{snapshot}/#{layer} not found", 404) if nw.nil?
-
-                ns_converter = ns_converter_wo_topology(network)
-                nodes = nw.nodes.map
-                nodes = nw.nodes.select { |n| n.attribute.node_type == params[:node_type] } if params.key?(:node_type)
-
-                # response
-                nodes.map do |node|
-                  {
-                    node: node.name,
-                    alias: ns_converter.node_name_table.find_l1_alias(node.name),
-                    attribute: node.attribute.to_data
-                  }
-                end
-              end
-
-              desc 'Get all interfaces in the layer'
-              params do
-                optional :node_type, type: String, desc: 'Node type (segment/node/endpoint)'
-              end
-              get 'interfaces' do
-                network, snapshot, layer = %i[network snapshot layer].map { |key| params[key] }
-                nws = read_topology_instance(network, snapshot)
-                nw = nws.find_network(layer)
-                error!("#{network}/#{snapshot}/#{layer} not found", 404) if nw.nil?
-
-                ns_converter = ns_converter_wo_topology(network)
-                nodes = nw.nodes
-                nodes = nw.nodes.select { |n| n.attribute.node_type == params[:node_type] } if params.key?(:node_type)
-
-                # response
-                nodes.map do |node|
-                  {
-                    node: node.name,
-                    alias: ns_converter.node_name_table.find_l1_alias(node.name),
-                    attribute: node.attribute.to_data,
-                    interfaces: node.termination_points.map do |tp|
-                      {
-                        interface: tp.name,
-                        alias: ns_converter.tp_name_table.find_l1_alias(node.name, tp.name),
-                        attribute: tp.attribute.to_data
-                      }
-                    end
-                  }
-                end
-              end
-            end
-          end
-
-          resource 'converted_topology' do
-            desc 'Get namespace-convert-table to get converted topology'
-            get do
-              network, snapshot = %i[network snapshot].map { |key| params[key] }
-              ns_converter = ns_converter_wo_topology(network)
-              ns_converter.load_origin_topology(read_topology_file(network, snapshot))
-              # response
-              ns_converter.convert
-            end
-          end
-        end
-      end
-      # rubocop:enable Metrics/BlockLength
-    end
+    mount ApiRoute::Topologies
   end
-  # rubocop:enable Metrics/ClassLength
 end
